Support local codec server config settings
