+++ src/lib/models/event-history/index.ts
@@ -1,8 +1,3 @@
-import {
-  codecEndpoint,
-  includeCredentials,
-  passAccessToken,
-} from '$lib/stores/data-encoder-config';
 import type {
   EventAttributeKey,
   EventAttributesWithType,
@@ -18,11 +13,6 @@ import {
   decodePayloadAttributes,
 } from '$lib/utilities/decode-payload';
 import { formatDate } from '$lib/utilities/format-date';
-import {
-  getCodecEndpoint,
-  getCodecIncludeCredentials,
-  getCodecPassAccessToken,
-} from '$lib/utilities/get-codec';
 import { has } from '$lib/utilities/has';
 import { findAttributesAndKey } from '$lib/utilities/is-event-type';
 
@@ -35,36 +25,13 @@ export async function getEventAttributes(
   {
     convertWithCodec = convertPayloadToJsonWithCodec,
     decodeAttributes = decodePayloadAttributes,
-    encoderEndpoint = codecEndpoint,
-    codecPassAccessToken = passAccessToken,
-    codecIncludeCredentials = includeCredentials,
   }: DecodeFunctions = {},
 ): Promise<EventAttributesWithType<EventAttributeKey>> {
   const { key, attributes } = findAttributesAndKey(historyEvent);
-  // Use locally set endpoint over settings endpoint for testing purposes
-  const endpoint = getCodecEndpoint(settings, encoderEndpoint);
-  const passAccessToken = getCodecPassAccessToken(
-    settings,
-    codecPassAccessToken,
-  );
-  const includeCredentials = getCodecIncludeCredentials(
-    settings,
-    codecIncludeCredentials,
-  );
-  const _settings = {
-    ...settings,
-    codec: {
-      ...settings?.codec,
-      endpoint,
-      passAccessToken,
-      includeCredentials,
-    },
-  };
-
   const convertedAttributes = await convertWithCodec({
     attributes,
     namespace,
-    settings: _settings,
+    settings,
     accessToken,
   });
