+++ src/lib/utilities/decode-payload.ts
@@ -19,6 +19,11 @@ import type {
 import type { Optional, Replace, Settings } from '$lib/types/global';
 
 import { atob } from './atob';
+import {
+  getCodecEndpoint,
+  getCodecIncludeCredentials,
+  getCodecPassAccessToken,
+} from './get-codec';
 import { has } from './has';
 import { isObject } from './is';
 import { parseWithBigInt } from './parse-with-big-int';
@@ -175,7 +180,24 @@ export const decodeAllPotentialPayloadsWithCodec = async (
   settings: Settings = get(page).data.settings,
   accessToken: string = get(authUser).accessToken,
 ): Promise<EventAttribute | PotentiallyDecodable> => {
-  const decode = decodePayloadWithCodec(namespace, settings, accessToken);
+  const endpoint = getCodecEndpoint(settings);
+  const passAccessToken = getCodecPassAccessToken(settings);
+  const includeCredentials = getCodecIncludeCredentials(settings);
+  const settingsWithLocalConfig = {
+    ...settings,
+    codec: {
+      ...settings?.codec,
+      endpoint,
+      passAccessToken,
+      includeCredentials,
+    },
+  };
+
+  const decode = decodePayloadWithCodec(
+    namespace,
+    settingsWithLocalConfig,
+    accessToken,
+  );
   if (anyAttributes) {
     for (const key of Object.keys(anyAttributes)) {
       if (keyIs(key, 'payloads', 'encodedAttributes') && anyAttributes[key]) {
